refactor journey utils
