+++ src/service/impl/trips/trips.ts
@@ -11,7 +11,7 @@ import * as Boom from '@hapi/boom';
 import {
   extractServiceJourneyIds,
   generateSingleTripQueryString
-} from '../../../utils/journey-utils';
+} from './utils';
 import * as Trips from '../../../types/trips';
 
 export async function getTrips(
