+++ src/service/impl/trips/utils.ts
@@ -1,15 +1,10 @@
 import Joi from 'joi';
-import { TripPattern } from '@entur/sdk';
-import {
-  TripsQueryVariables,
-  TripsQuery
-} from '../service/impl/trips/journey-gql/trip.graphql-gen';
+import { TripsQueryVariables } from './journey-gql/trip.graphql-gen';
 import {
   Leg,
   TripPattern as TripPattern_v3,
   TripsQueryWithJourneyIds
-} from '../types/trips';
-import { TripPatternsQuery, TripPatternQuery } from '../service/types';
+} from '../../../types/trips';
 import {
   compressToEncodedURIComponent,
   decompressFromEncodedURIComponent
@@ -17,50 +12,6 @@ import {
 import { addSeconds, parseISO } from 'date-fns';
 
 const START_TIME_PADDING = 60; // time in seconds
-export function generateId(trip: TripPattern, query: TripPatternsQuery) {
-  const fields: TripPatternsQuery = {
-    searchDate: new Date(trip.startTime),
-    ...query
-  };
-  const serviceIds = getServiceIds(trip);
-  return compressToEncodedURIComponent(
-    JSON.stringify({ query: fields, serviceIds })
-  );
-}
-
-export function getServiceIds(trip: TripPattern) {
-  return trip.legs.map(leg => leg.serviceJourney?.id ?? 'null');
-}
-
-export function parseTripPatternId(
-  id: string,
-  queryValidator: Joi.ObjectSchema<any>
-): TripPatternQuery {
-  try {
-    const value = decompressFromEncodedURIComponent(id);
-    if (!value) {
-      throw new Error();
-    }
-    const fields = JSON.parse(value);
-    fields.query = queryValidator.validate(fields.query).value;
-
-    if (isTripPatternsQuery(fields.query) && Array.isArray(fields.serviceIds)) {
-      return fields as TripPatternQuery;
-    }
-  } catch (_) {}
-
-  throw new Error('Could not parse input trip id');
-}
-
-function isTripPatternsQuery(
-  potentialTripPatternsQuery: any
-): potentialTripPatternsQuery is TripPatternsQuery {
-  return (
-    'from' in potentialTripPatternsQuery &&
-    'searchDate' in potentialTripPatternsQuery &&
-    'modes' in potentialTripPatternsQuery
-  );
-}
 
 /**
  * Creates a unique query to fetch updates to a single Trip
